backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0
bcrypt>=4.0.0
python-jose[cryptography]
python-multipart
certifi